Split listener.rs into listener/ submodules
Extract listener.rs (3594 lines) into listener/ directory with submodules:
- listener/state.rs: ListenerState, ListenerConfig, constants
- listener/forwarding.rs: KeyEventForwarder trait, UinputForwarder
- listener/sequence.rs: SequenceRuntime, ActiveSequence, SequenceDispatch
- listener/device.rs: DeviceState, ModifierTracker
- listener/hotplug.rs: RawFdGuard, hotplug inotify event parsing
- listener/dispatch.rs: callback dispatch/collection functions
- listener/io.rs: thread spawning, device open/poll/read, key state

listener.rs retains: module declarations, listener_loop, and all tests.

Also clean up unused imports exposed during the split in manager.rs,
backend.rs, and mode/ submodules.

All 219 tests pass (--all-features), zero warnings, clean clippy.
